+++ DataverseToPowerBI.XrmToolBox/Properties/AssemblyInfo.cs
@@ -15,6 +15,6 @@ using System.Runtime.InteropServices;
 
 [assembly: Guid("b8c9a3e4-2d6f-4a8b-9e5c-1f7d3a6b8c9e")]
 
-[assembly: AssemblyVersion("1.2026.4.16")]
-[assembly: AssemblyFileVersion("1.2026.4.16")]
+[assembly: AssemblyVersion("1.2026.5.4")]
+[assembly: AssemblyFileVersion("1.2026.5.4")]